backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.7.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
billiard==4.2.1
blinker==1.8.2
cachelib==0.9.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
dnspython==2.6.1
email_validator==2.2.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.10
langchain-community==0.3.10
langchain-core==0.3.22
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2
langsmith==0.1.147
MarkupSafe==2.1.5
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.3
orjson==3.10.12
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.6
WTForms==3.1.2
yarl==1.18.3
